feat(胆固醇): 添加胆固醇记录功能及相关界面
实现胆固醇记录功能，包括：
1. 新增胆固醇记录页面及交互逻辑
2. 修改数据类型从Float到Int
3. 添加胆固醇计算工具类
4. 实现胆固醇等级评估功能
5. 添加相关字符串资源和布局文件

+++ app/src/main/java/com/healthtracker/blood/suger/data/entity/CholesterolRecord.kt
@@ -23,15 +23,15 @@ data class CholesterolRecord(
 
     /** 高密度脂蛋白 HDL（mg/dL） */
     @ColumnInfo(name = "hdl")
-    val hdl: Float? = null,
+    val hdl: Int? = null,
 
     /** 低密度脂蛋白 LDL（mg/dL） */
     @ColumnInfo(name = "ldl")
-    val ldl: Float? = null,
+    val ldl: Int? = null,
 
     /** 甘油三酯 TG（mg/dL） */
     @ColumnInfo(name = "triglyceride")
-    val triglyceride: Float? = null,
+    val triglyceride: Int? = null,
 
     /** 总胆固醇 TC（mg/dL） */
     @ColumnInfo(name = "tc")
